+++ src/widgets/index.ts
@@ -29,4 +29,4 @@ export { default as queryRuleContext } from './query-rule-context/query-rule-con
 export { default as index } from './index/index';
 export { default as places } from './places/places';
 export { default as EXPERIMENTAL_answers } from './answers/answers';
-export { default as smartSort } from './smart-sort/smart-sort';
+export { default as relevantSort } from './relevant-sort/relevant-sort';