+++ src/widgets/toggle-refinement/__tests__/toggle-refinement-test.js
@@ -5,7 +5,7 @@ import RefinementList from '../../../components/RefinementList/RefinementList';
 import { createInstantSearch } from '../../../../test/mock/createInstantSearch';
 
 jest.mock('preact', () => {
-  const module = require.requireActual('preact');
+  const module = jest.requireActual('preact');
 
   module.render = jest.fn();