+++ test/utils/castToJestMock.ts
@@ -1,1 +1,3 @@
-export const castToJestMock = (obj: any): jest.Mock => obj;
+export const castToJestMock = <TFunction extends (...args: any[]) => any>(
+  func: TFunction
+) => func as jest.MockedFunction<typeof func>;